refactor: move `EVMSortMergeJoinExec` to its own file
